move filter to TuistGraph.Graph extension
