+++ Sources/TuistKit/Services/GraphService.swift
@@ -54,8 +54,7 @@ final class GraphService {
             try FileHandler.shared.delete(filePath)
         }
 
-        let filteredTargetsAndDependencies = filter(
-            graph: graph,
+        let filteredTargetsAndDependencies = graph.filter(
             skipTestTargets: skipTestTargets,
             skipExternalDependencies: skipExternalDependencies,
             targetsToFilter: targetsToFilter
@@ -78,52 +77,6 @@ final class GraphService {
         }
 
         logger.notice("Graph exported to \(filePath.pathString)", metadata: .success)
-    }
-
-    /// Filters the project graph
-    /// - Parameters:
-    ///   - graph: Graph to be filtered
-    /// - Returns: Filtered graph targets and dependencies
-    private func filter(
-        graph: TuistGraph.Graph,
-        skipTestTargets: Bool,
-        skipExternalDependencies: Bool,
-        targetsToFilter: [String]
-    ) -> [GraphTarget: Set<GraphDependency>] {
-        let graphTraverser = GraphTraverser(graph: graph)
-
-        let allTargets: Set<GraphTarget> = skipExternalDependencies ? graphTraverser.allInternalTargets() : graphTraverser
-            .allTargets()
-        let filteredTargets: Set<GraphTarget> = allTargets.filter { target in
-            if skipTestTargets, graphTraverser.dependsOnXCTest(path: target.path, name: target.target.name) {
-                return false
-            }
-
-            if !targetsToFilter.isEmpty, !targetsToFilter.contains(target.target.name) {
-                return false
-            }
-
-            return true
-        }
-
-        let filteredTargetsAndDependencies: Set<GraphTarget> = filteredTargets.union(
-            transitiveClosure(Array(filteredTargets)) { target in
-                Array(graphTraverser.directTargetDependencies(path: target.path, name: target.target.name))
-            }
-        )
-
-        return filteredTargetsAndDependencies.reduce(into: [GraphTarget: Set<GraphDependency>]()) { result, target in
-            if skipExternalDependencies, target.project.isExternal { return }
-
-            guard let targetDependencies = graphTraverser.dependencies[.target(name: target.target.name, path: target.path)]
-            else { return }
-
-            result[target] = targetDependencies
-                .filter { dependency in
-                    if skipExternalDependencies, dependency.isExternal(graph.projects) { return false }
-                    return true
-                }
-        }
     }
 
     private func export(
@@ -272,16 +225,5 @@ extension ProjectAutomation.Scheme {
         }
 
         return ProjectAutomation.Scheme(name: scheme.name, testActionTargets: testTargets)
-    }
-}
-
-extension GraphDependency {
-    fileprivate func isExternal(_ projects: [AbsolutePath: TuistGraph.Project]) -> Bool {
-        switch self {
-        case let .target(_, path):
-            return projects[path]?.isExternal ?? false
-        case .framework, .xcframework, .library, .bundle, .packageProduct, .sdk:
-            return true
-        }
     }
 }
